Update quiz.py
added loop code that goes through each quiz question one by one. It adds 1 to the loop index to get the question number and then prints the question text from the questions list. This shows each question in order with the correct number for the player. At the moment it just prints all questions but with more code it will work.