Update quiz.py
I had a few people test the quiz and I realised it was too easy to pass so I changed it to 13 correct answers to pass from 10. Added some emojis to improve appearance. Added code that adds a bonus checkpoint every 5 questions. It uses question_number % 5 == 0 to check if the question number is a multiple of 5, and also makes sure it’s not the last question. If both conditions are true, it prints a message to encourage the player to keep going.