Update quiz.py
Improvement After Testing
I had Alex test my quiz and he suggested an option to play the quiz again at the end. So I added the question "Do you want to play the quiz again?" and if the user enters "yes" the program restarts. If they enter "no" it prints a goodbye message and program ends.