Update quiz.py
Improvement After Testing.
Had two more people test my quiz and they both entered what they though the answer was instead of a/b/c/d on the multi-choice. So, I changed it to ask the user to Enter a/b/c/d. I did this to the true/false question as well just in case. I have tested this and it works.