Update quiz.py
Tested the hard quiz and it worked great but I found there was a repeat of a question so I replaced it. Changed the do you want to play the quiz do you want to play a quiz since there is two now. Changed the questions message to just hello so it is not being said twice. Changed the other hello to great choice so it is not being said twice.