Update quiz.py
Improvement after testing. I noticed that people who played constantly missed a message. 'There are 20 questions and you need 10 correct answers to pass'. So, I added a pause so they have time to read it and don't miss it, then added a countdown into the quiz so they are ready.